requirements.txt
appdirs==1.4.3
asn1crypto==0.23.0
astroid==1.5.3
awscli==1.16.105
Babel==2.5.0
beautifulsoup4==4.6.0
boto3==1.9.95
botocore==1.12.95
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
Collectfast==0.6.2
colorama==0.3.7
cryptography==2.1.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-appconf==1.0.2
django-bootstrap3==8.2.3
django-debug-toolbar==1.9.1
django-extensions==1.9.8
django-extra-fields==1.0.0
django-filter==1.1.0
django-heroku==0.3.1
django-imagekit==4.0.1
django-localflavor==1.4.1
django-material==1.0.0
django-mathfilters==0.4.0
django-mptt==0.8.7
django-phonenumber-field==1.3.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
django-rq==2.0
django-storages==1.7.1
django-widget-tweaks==1.4.1
docutils==0.13.1
google-api-core==0.1.2
google-auth==1.2.1
google-cloud-language==1.0.0
googleapis-common-protos==1.5.3
googlemaps==2.5.1
grpcio==1.7.3
gunicorn==19.7.1
idna==2.6
isort==4.2.15
jmespath==0.9.2
lazy-object-proxy==1.3.1
markdown==2.6.11
mccabe==0.6.1
mock==2.0.0
nexmo==1.5.0
nltk==3.2.4
numpy==1.13.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pbr==3.1.1
phonenumberslite==8.8.1
pilkit==2.0
Pillow==4.1.1
protobuf==3.5.0.post1
pstat==0.0.1
psutil==5.4.2
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyJWT==1.5.3
pylint==1.7.4
pyparsing==2.2.0
python-dateutil==2.6.0
python-socketio==4.2.0
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.21.0
requests-oauthlib==0.8.0
redis>=3
rq>=0.13
rsa==3.4.2
s3transfer==0.2.0
six==1.11.0
sqlparse==0.2.4
stripe==2.25.0
typing==3.6.2
urllib3==1.22
whitenoise==3.3.0
wrapt==1.10.11
